OMS - 供应链 - 实仓调整

+++ oms-modules/oms-supplychain/src/main/java/com/oms/supplychain/model/entity/warehouse/WmsRealStoreInfo.java
@@ -107,7 +107,7 @@ public class WmsRealStoreInfo implements Serializable {
      * 1 真实出库 2 虚拟出库
      */
     @Excel(name = "出入库类型")
-    private Byte actualWarehouse;
+    private Integer actualWarehouse;
 
     /**
      * 货主
